refactor, try prevent infinite rerender hot-reload issue
