+++ listen-interface/src/i18n/translations/pipelines.ts
@@ -25,8 +25,8 @@ export const pipelines = {
       Failed: "Failed",
       Cancelled: "Cancelled",
     },
-    approve: "Approve",
-    reject: "Reject",
+    approve: "Confirm",
+    reject: "Cancel",
     pipeline_rejected: "Order rejected",
     pipeline_scheduled_for_execution: "Order scheduled for execution",
     step_cancelled: "Step cancelled",
